// app/[locale]/types.ts
export type Locale =
  | "en"
  | "ru"
  | "de"
  | "es"
  | "it"
  | "pt-br"
  | "fr"
  | "pl"
  | "cs"
  | "tr"
  | "ja"
  | "th"
  | "zh-cn"
  | "zh-tw"
  | "ko";

export const SUPPORTED_LOCALES: Locale[] = [
  "en",
  "ru",
  "de",
  "es",
  "it",
  "pt-br",
  "fr",
  "pl",
  "cs",
  "tr",
  "ja",
  "th",
  "zh-cn",
  "zh-tw",
  "ko",
];